Add integration test for /health_check
