+++ src/main.rs
@@ -1,24 +1,6 @@
-use actix_web::{web, App, HttpRequest, HttpServer, Responder, HttpResponse};
-
-async fn greet(req: HttpRequest) -> impl Responder {
-    let name = req.match_info().get("name").unwrap_or("World");
-    format!("Hello {}!", &name)
-}
-
-async fn health_check() -> impl Responder {
-    HttpResponse::Ok()
-}
+use zero2prod::run;
 
 #[tokio::main]
 async fn main() -> std::io::Result<()> {
-    HttpServer::new(|| {
-        App::new()
-            .route("/", web::get().to(greet))
-            .route("/health_check", web::get().to(health_check))
-            .route("/{name}", web::get().to(greet))
-
-    })
-    .bind("127.0.0.1:8000")?
-        .run()
-        .await
+    run()?.await
 }
